shopping_cart_backend/requirements.txt
django==3.0
djangorestframework==3.11.0
